Moved buffer management logic into nitf module.  Moved core byte manipulation into base ByteProvider class
